+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Master
 
+* Fix a bug where `SomeClass.any_instance.should_receive(:message).any_number_of_times` was converted into `expect_any_instance_of(SomeClass).to receive(:message)` unintentionally (now it's converted into `allow_any_instance_of(SomeClass).to receive(:message)`)
+
 ## v0.0.5
 
 * Support conversion of `any_number_of_times`